requirements.txt
apispec==3.3.0
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
decorators==0.1.1
defusedxml==0.6.0
dsnparse==0.1.15
enum34==1.1.10
Flask==1.1.2
Flask-Babel==1.0.0
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.1
idna==2.9
IMDbPY==6.8
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
JustWatch==0.5.1
lxml==4.5.0
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-enum==1.5.1
pkg-resources==0.0.0
prison==0.1.3
pyrsistent==0.16.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
redis==3.4.1
redis-collections==0.8.0
requests==2.23.0
rotten-tomatoes-client==0.0.3
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.16
SQLAlchemy-Utils==0.36.3
tmdb3==0.7.2
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.2.1
zipp==3.1.0
